core/standards/scripts/scheduler2/location_hash/404.html

<!doctype html>
<title>Location hash - set hash then go back in history and set hash in many code blocks</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
var step_delay = 100;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});

window.onload = function () {
    setTimeout(step1, step_delay);  
};

function step1() {
    t.step(function() {
        location.hash = '#foo';
    });
    setTimeout(step2, step_delay);
}

function step2() {
    t.step(function() {
        history.go(-1);
    });
    setTimeout(step3, step_delay);
}

function step3() {
    t.step(function() {
        location.hash = '#bar';
    });
    setTimeout(last_step, step_delay);
}

function last_step() {
    t.step(function() {
        assert_equals( location.hash, '#bar' );
        assert_equals( history.length, histlength + 1 );
    });
    t.done();
}
</script>
<div id="log"></div>
